refactor: Reorginized components and elements
Moved player, meta, primaries, secondaries into components folder
