+++ src/elements/numInput.svelte
@@ -4,7 +4,7 @@
   export let passedClasses = null;
   export let border = false;
 
-  import { players, activePlayer } from "../stores.js";
+  import { players, activePlayer } from "../stores";
 </script>
 
 <style>
